+++ ME0StubFinder/interface/Subfunc.h
@@ -58,14 +58,14 @@ public:
 
 hi_lo_t mirror_hi_lo(const hi_lo_t& ly);
 patdef_t mirror_patdef(const patdef_t& pat, int id);
-std::vector<hi_lo_t> create_pat_ly(float lower, float upper);
+std::vector<hi_lo_t> create_pat_ly(double lower, double upper);
 
 int count_ones(uint64_t x);
 UInt192 set_bit(int index, UInt192& num1);
 UInt192 clear_bit(int index, UInt192 num);
 uint64_t ones_bit_mask(int num);
 std::vector<int> find_ones(uint64_t& data);
-float find_centroid(uint64_t& data);
+double find_centroid(uint64_t& data);
 std::vector<std::vector<ME0Stub>> chunk(const std::vector<ME0Stub>& in_list, int n);
 void segment_sorter(std::vector<ME0Stub>& segs, int n); 
 std::vector<ME0Stub> concatVector(const std::vector<std::vector<ME0Stub>>& vec);
